Move builtins to a separate module
