+++ src/main.rs
@@ -2,11 +2,10 @@ use std::collections::BTreeMap;
 use std::string::String;
 use std::vec::Vec;
 use std::boxed::Box;
-use std::fs::{self, File};
+use std::fs::File;
 use std::io::{stdout, Read, Write};
 use std::env;
 use std::process;
-use std::thread;
 
 use self::to_num::ToNum;
 use self::input_editor::readln;
@@ -14,6 +13,7 @@ use self::tokenizer::{Token, tokenize};
 use self::expansion::expand_tokens;
 use self::parser::{parse, Job};
 
+pub mod builtin;
 pub mod to_num;
 pub mod input_editor;
 pub mod tokenizer;
@@ -46,227 +46,6 @@ pub struct Shell<'a> {
     pub mode: &'a mut Vec<Mode>,
 }
 
-pub fn builtin_cat(args: &Vec<String>) {
-    let path = args.get(1).map_or(String::new(), |arg| arg.clone());
-
-    match File::open(&path) {
-        Ok(mut file) => {
-            let mut string = String::new();
-            match file.read_to_string(&mut string) {
-                Ok(_) => println!("{}", string),
-                Err(err) => {
-                    println!("Failed to read: {}: {}", path, err)
-                }
-            }
-        }
-        Err(err) => println!("Failed to open file: {}: {}", path, err),
-    }
-}
-
-pub fn builtin_cd(args: &Vec<String>) {
-    match args.get(1) {
-        Some(path) => {
-            if let Err(err) = env::set_current_dir(&path) {
-                println!("Failed to set current dir to {}: {}",
-                         path,
-                         err);
-            }
-        }
-        None => println!("No path given"),
-    }
-}
-
-pub fn builtin_echo(args: &Vec<String>) {
-    let echo = args.iter()
-        .skip(1)
-        .fold(String::new(),
-        |string, arg| string + " " + arg);
-    println!("{}", echo.trim());
-}
-
-pub fn buiiltin_free() {
-    match File::open("memory:") {
-        Ok(mut file) => {
-            let mut string = String::new();
-            match file.read_to_string(&mut string) {
-                Ok(_) => println!("{}", string),
-                Err(err) => println!("Failed to read: memory: {}", err),
-            }
-        }
-        Err(err) => println!("Failed to open file: memory: {}", err),
-    }
-}
-
-pub fn builtin_ls(args: &Vec<String>) {
-    let path = args.get(1).map_or(".".to_string(), |arg| arg.clone());
-
-    let mut entries = Vec::new();
-    match fs::read_dir(&path) {
-        Ok(dir) => {
-            for entry_result in dir {
-                match entry_result {
-                    Ok(entry) => {
-                        let directory = match entry.file_type() {
-                            Ok(file_type) => file_type.is_dir(),
-                            Err(err) => {
-                                println!("Failed to read file type: {}", err);
-                                false
-                            }
-                        };
-
-                        match entry.file_name().to_str() {
-                            Some(path_str) => {
-                                if directory {
-                                    entries.push(path_str.to_string() + "/")
-                                } else {
-                                    entries.push(path_str.to_string())
-                                }
-                            }
-                            None => {
-                                println!("Failed to convert path to string")
-                            }
-                        }
-                    }
-                    Err(err) => {
-                        println!("Failed to read entry: {}", err)
-                    }
-                }
-            }
-        }
-        Err(err) => {
-            println!("Failed to open directory: {}: {}", path, err)
-        }
-    }
-
-    entries.sort();
-
-    for entry in entries {
-        println!("{}", entry);
-    }
-}
-
-pub fn builtin_mkdir(args: &Vec<String>) {
-    match args.get(1) {
-        Some(dir_name) => {
-            if let Err(err) = fs::create_dir(dir_name) {
-                println!("Failed to create: {}: {}", dir_name, err);
-            }
-        }
-        None => println!("No name provided"),
-    }
-}
-
-pub fn builtin_poweroff() {
-    match File::create("acpi:off") {
-        Err(err) => println!("Failed to remove power (error: {})", err),
-        Ok(_) => println!("I see dead people"),
-    }
-}
-
-pub fn builtin_ps() {
-    match File::open("context:") {
-        Ok(mut file) => {
-            let mut string = String::new();
-            match file.read_to_string(&mut string) {
-                Ok(_) => println!("{}", string),
-                Err(err) => {
-                    println!("Failed to read: context: {}", err)
-                }
-            }
-        }
-        Err(err) => println!("Failed to open file: context: {}", err),
-    }
-}
-
-pub fn builtin_pwd() {
-    match env::current_dir() {
-        Ok(path) => {
-            match path.to_str() {
-                Some(path_str) => println!("{}", path_str),
-                None => println!("?"),
-            }
-        }
-        Err(err) => println!("Failed to get current dir: {}", err),
-    }
-}
-
-pub fn builtin_read(args: &Vec<String>, variables: &mut BTreeMap<String, String>) {
-    for i in 1..args.len() {
-        if let Some(arg_original) = args.get(i) {
-            let arg = arg_original.trim();
-            print!("{}=", arg);
-            if let Err(message) = stdout().flush() {
-                println!("{}: Failed to flush stdout", message);
-            }
-            if let Some(value_original) = readln() {
-                let value = value_original.trim();
-                set_var(variables, arg, value);
-            }
-        }
-    }
-}
-
-pub fn builtin_rm(args: &Vec<String>) {
-    match args.get(1) {
-        Some(path) => {
-            if fs::remove_file(path).is_err() {
-                println!("Failed to remove: {}", path);
-            }
-        }
-        None => println!("No name provided"),
-    }
-}
-
-pub fn builtin_rmdir(args: &Vec<String>) {
-    match args.get(1) {
-        Some(path) => {
-            if fs::remove_dir(path).is_err() {
-                println!("Failed to remove: {}", path);
-            }
-        }
-        None => println!("No name provided"),
-    }
-}
-
-pub fn builtin_run(args: &Vec<String>, variables: &mut BTreeMap<String, String>) {
-    let path = "/apps/shell/main.bin";
-
-    let mut command = process::Command::new(path);
-    for i in 1..args.len() {
-        if let Some(arg) = args.get(i) {
-            command.arg(arg);
-        }
-    }
-
-    match command.spawn() {
-        Ok(mut child) => {
-            match child.wait() {
-                Ok(status) => {
-                    if let Some(code) = status.code() {
-                        set_var(variables, "?", &format!("{}", code));
-                    } else {
-                        println!("{}: No child exit code", path);
-                    }
-                }
-                Err(err) => {
-                    println!("{}: Failed to wait: {}", path, err)
-                }
-            }
-        }
-        Err(err) => println!("{}: Failed to execute: {}", path, err),
-    }
-}
-
-pub fn builtin_sleep(args: &Vec<String>) {
-    let secs = args.get(1).map_or(0, |arg| arg.to_num());
-    thread::sleep_ms(secs as u32 * 1000);
-}
-
-pub fn builtin_touch(args: &Vec<String>) {
-    let secs = args.get(1).map_or(0, |arg| arg.to_num());
-    thread::sleep_ms(secs as u32 * 1000);
-}
-
 impl Command {
     /// Return the map from command names to commands
     pub fn map() -> BTreeMap<String, Self> {
@@ -279,7 +58,7 @@ impl Command {
                             main: Box::new(|args: &Vec<String>,
                                             _: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                builtin_cat(args);
+                                builtin::cat(args);
                             }),
                         });
 
@@ -290,7 +69,7 @@ impl Command {
                             main: Box::new(|args: &Vec<String>,
                                             _: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                builtin_cd(args);
+                                builtin::cd(args);
                             }),
                         });
 
@@ -301,7 +80,7 @@ impl Command {
                             main: Box::new(|args: &Vec<String>,
                                             _: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                builtin_echo(args);
+                                builtin::echo(args);
                             }),
                         });
 
@@ -321,7 +100,7 @@ impl Command {
                             main: Box::new(|_: &Vec<String>,
                                             _: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                buiiltin_free();
+                                builtin::free();
                             }),
                         });
 
@@ -332,7 +111,7 @@ impl Command {
                             main: Box::new(|args: &Vec<String>,
                                             _: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                builtin_ls(args);
+                                builtin::ls(args);
                             }),
                         });
 
@@ -344,7 +123,7 @@ impl Command {
                             main: Box::new(|args: &Vec<String>,
                                             _: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                builtin_mkdir(args);
+                                builtin::mkdir(args);
                             }),
                         });
 
@@ -356,7 +135,7 @@ impl Command {
                             main: Box::new(|_: &Vec<String>,
                                             _: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                builtin_poweroff();
+                                builtin::poweroff();
                             }),
                         });
 
@@ -367,7 +146,7 @@ impl Command {
                             main: Box::new(|_: &Vec<String>,
                                             _: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                builtin_ps();
+                                builtin::ps();
                             }),
                         });
 
@@ -378,7 +157,7 @@ impl Command {
                             main: Box::new(|_: &Vec<String>,
                                             _: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                builtin_pwd();
+                                builtin::pwd();
                             }),
                         });
 
@@ -389,7 +168,7 @@ impl Command {
                             main: Box::new(|args: &Vec<String>,
                                             variables: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                builtin_read(args, variables);
+                                builtin::read(args, variables);
                             }),
                         });
 
@@ -400,7 +179,7 @@ impl Command {
                             main: Box::new(|args: &Vec<String>,
                                             _: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                builtin_rm(args);
+                                builtin::rm(args);
                             }),
                         });
 
@@ -411,7 +190,7 @@ impl Command {
                             main: Box::new(|args: &Vec<String>,
                                             _: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                builtin_rmdir(args);
+                                builtin::rmdir(args);
                             }),
                         });
 
@@ -422,7 +201,7 @@ impl Command {
                             main: Box::new(|args: &Vec<String>,
                                             variables: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                builtin_run(args, variables);
+                                builtin::run(args, variables);
                             }),
                         });
 
@@ -434,7 +213,7 @@ impl Command {
                             main: Box::new(|args: &Vec<String>,
                                             _: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                builtin_sleep(args);
+                                builtin::sleep(args);
                             }),
                         });
 
@@ -448,7 +227,7 @@ impl Command {
                             main: Box::new(|args: &Vec<String>,
                                             _: &mut BTreeMap<String, String>,
                                             _: &mut Vec<Mode>| {
-                                builtin_touch(args);
+                                builtin::touch(args);
                             }),
                         });
